recomend test update

+++ SSM-Demo/bamboo/src/main/java/com/bamboo/dto/RoomPreferencesDTO.java
@@ -14,5 +14,5 @@ public class RoomPreferencesDTO {
     private String liveCode;
 
     @NotEmpty(message = "preference不能为空")
-    private Integer preference;
+    private Float preference;
 }
